third_party/blink/web_tests/fast/forms/select-popup/popup-menu-size.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/picker-common.js"></script>
<style>
#select1 {
    -webkit-appearance: none;
    width: 100px;
    padding: 1px 30px 1px 6px;
}
</style>
<select id="select1">
<option>abc def ghi</option>
<option>o2</option>
<option>o3</option>
<option>o4</option>
<option>o5</option>
<option>o6</option>
<option>o7</option>
<option>o8</option>
<option>o9</option>
<option>o10</option>
<option>o11</option>
<option>o12</option>
<option>o13</option>
<option>o14</option>
<option>o15</option>
<option>o16</option>
<option>o17</option>
<option>o18</option>
<option>o19</option>
<option>o20</option>
<option>o21</option>
</select>

<script>
function openPickerErrorCallback() {
    testFailed('picker didn\'t open')
    finishJSTest();
}

var select1 = document.getElementById('select1');

openPickerDeprecatedJsTest(select1, function() {
    shouldBe('internals.pagePopupWindow.innerWidth', '100');
    finishJSTest();
}, openPickerErrorCallback);
</script>
